refactor: 프로젝트 생성 및 수정 API 로직 분리
