+++ features/project/api/create-project-form-data.ts
@@ -1,0 +1,10 @@
+import type { ProjectRequest } from '../types';
+
+export const createProjectFormData = (request: ProjectRequest, image?: File | null): FormData => {
+  const formData = new FormData();
+
+  formData.append('request', new Blob([JSON.stringify(request)], { type: 'application/json' }));
+  formData.append('image', image ?? '');
+
+  return formData;
+};
