+++ features/project/index.ts
@@ -3,11 +3,9 @@ export { EditProjectModal } from './components/edit-project-modal';
 export { ProjectsPageContent } from './components/projects-page-content';
 export {
   useConfirmProjectMembers,
-  useCreateProject,
   useDeleteProject,
   useDeleteProjectMember,
   useInviteProjectMember,
-  useUpdateProject,
 } from './api/mutations';
 export { fetchProjects } from './api/fetchers';
 export { useProject, useProjects, useUserProjects } from './api/queries';
@@ -17,6 +15,8 @@ export { ProjectSection } from './components/project-section';
 export { ProjectSectionStream } from './components/project-section-stream';
 export { ProjectSectionSkeleton } from './components/project-section.skeleton';
 export { useProjectCards } from './hooks/use-project-cards';
+export { useCreateProject } from './hooks/use-create-project';
+export { useUpdateProject } from './hooks/use-update-project';
 export { mapMyProjectToCardItem, mapMyProjectsToCardItems } from './lib/map-my-project-to-card';
 export type {
   CreateProjectPayload,
