--- features/project/lib/build-project-multipart.ts
@@ -1,13 +1,0 @@
-import type { ProjectCreateRequest, ProjectUpdateRequest } from '../types';
-
-export const buildProjectMultipartFormData = (
-  request: ProjectCreateRequest | ProjectUpdateRequest,
-  image?: File | null,
-): FormData => {
-  const formData = new FormData();
-  formData.append('request', new Blob([JSON.stringify(request)], { type: 'application/json' }));
-  if (image) {
-    formData.append('image', image);
-  }
-  return formData;
-};
